Created all exceptions.

+++ src/main/java/com/tt/test/service/impl/AbilityServiceImpl.java
@@ -7,6 +7,7 @@ import com.tt.test.repository.EmployeeEntityRepository;
 import com.tt.test.service.AbilityService;
 import com.tt.test.service.dto.AbilityDTO;
 import com.tt.test.service.mapper.AbilityMapper;
+import com.tt.test.web.rest.exceptions.ResourceNotFoundException;
 import fr.xebia.extras.selma.Selma;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -28,15 +29,12 @@ public class AbilityServiceImpl implements AbilityService {
     }
 
     @Override
-    // na razie zostaw bo potrzebne do tworzenia w database init. Potem sie wywali
     public void create(AbilityEntity obj) {
         abilityEntityRepository.save(obj);
     }
 
     @Override
     public AbilityEntity create(AbilityDTO abilityDTO) {
-
-        //sprawdz czy istnieje taki o podanym id
         EmployeeEntity employeeById = findEmployeeById(abilityDTO.getEmployeeId());
         AbilityEntity abilityEntity = abilityMapper.asAbilityEntity(abilityDTO);
         abilityEntity.setEmployeeEntity(employeeById);
@@ -47,12 +45,18 @@ public class AbilityServiceImpl implements AbilityService {
 
     @Override
     public EmployeeEntity findEmployeeById(Long id) {
-        return employeeEntityRepository.findOne(id);
+        EmployeeEntity employeeEntity = employeeEntityRepository.findOne(id);
+        if (employeeEntity == null)
+            throw new ResourceNotFoundException("Employee with id = " + id + " was not found.");
+        return employeeEntity;
     }
 
     @Override
     public AbilityEntity getAbilityById(Long id) {
-        return abilityEntityRepository.findOne(id);
+        AbilityEntity abilityEntity = abilityEntityRepository.findOne(id);
+        if (abilityEntity == null)
+            throw new ResourceNotFoundException("Ability with id = " + id + " was not found.");
+        return abilityEntity;
     }
 
     @Override
@@ -62,10 +66,9 @@ public class AbilityServiceImpl implements AbilityService {
 
     @Override
     public AbilityEntity updateAbility(Long id, AbilityDTO abilityDTO) {
-        //sprawdz czy istnieje
-        AbilityEntity abilityById = getAbilityById(id);
         AbilityEntity abilityEntity = abilityMapper.asAbilityEntity(abilityDTO);
-        abilityEntity.setId(abilityById.getId());
+        if(checkIfExists(id))
+            abilityEntity.setId(id);
 
         EmployeeEntity employeeById = findEmployeeById(abilityDTO.getEmployeeId());
         abilityEntity.setEmployeeEntity(employeeById);
@@ -75,6 +78,14 @@ public class AbilityServiceImpl implements AbilityService {
 
     @Override
     public void deleteAbilityById(Long id) {
-        abilityEntityRepository.delete(id);
+        if(checkIfExists(id))
+            abilityEntityRepository.delete(id);
+    }
+
+    @Override
+    public boolean checkIfExists(Long id) {
+        if (!abilityEntityRepository.exists(id))
+            throw new ResourceNotFoundException("Ability with id = " + id + " was not found.");
+        return true;
     }
 }
