+++ src/main/java/com/tt/test/service/impl/PermissionServiceImpl.java
@@ -7,6 +7,7 @@ import com.tt.test.repository.PermissionEntityRepository;
 import com.tt.test.service.PermissionService;
 import com.tt.test.service.dto.PermissionDTO;
 import com.tt.test.service.mapper.PermissionMapper;
+import com.tt.test.web.rest.exceptions.ResourceNotFoundException;
 import fr.xebia.extras.selma.Selma;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -44,12 +45,18 @@ public class PermissionServiceImpl implements PermissionService {
 
     @Override
     public EmployeeEntity findEmployeeById(Long id) {
-        return employeeEntityRepository.findOne(id);
+        EmployeeEntity employeeEntity = employeeEntityRepository.findOne(id);
+        if (employeeEntity == null)
+            throw new ResourceNotFoundException("Employee with id = " + id + " was not found.");
+        return employeeEntity;
     }
 
     @Override
     public PermissionEntity getPermissionById(Long id) {
-        return permissionEntityRepository.findOne(id);
+        PermissionEntity permissionEntity = permissionEntityRepository.findOne(id);
+        if (permissionEntity == null)
+            throw new ResourceNotFoundException("Permission with id = " + id + " was not found.");
+        return permissionEntity;
     }
 
     @Override
@@ -59,10 +66,9 @@ public class PermissionServiceImpl implements PermissionService {
 
     @Override
     public PermissionEntity updatePermission(Long id, PermissionDTO permissionDTO) {
-        //sprawdz czy istnieje
-        PermissionEntity permissionById = getPermissionById(id);
         PermissionEntity permissionEntity = permissionMapper.asPermissionEntity(permissionDTO);
-        permissionEntity.setId(permissionById.getId());
+        if (checkIfExists(id))
+            permissionEntity.setId(id);
 
         EmployeeEntity employeeById = findEmployeeById(permissionDTO.getEmployeeId());
         permissionEntity.setEmployeeEntity(employeeById);
@@ -72,6 +78,14 @@ public class PermissionServiceImpl implements PermissionService {
 
     @Override
     public void deletePermissionById(Long id) {
-        permissionEntityRepository.delete(id);
+        if (checkIfExists(id))
+            permissionEntityRepository.delete(id);
+    }
+
+    @Override
+    public boolean checkIfExists(Long id) {
+        if (!permissionEntityRepository.exists(id))
+            throw new ResourceNotFoundException("Permission with id = " + id + " was not found.");
+        return true;
     }
 }
